+++ src/components/admin/Customers.tsx
@@ -13,7 +13,8 @@ import { RootState, AppDispatch } from "../../store";
 import { toast } from "react-hot-toast";
 import { customerService, Customer } from "../../services/customerService";
 import { useHeader } from "../../context/HeaderContext";
-import { policyService, Policy } from "../../services/policyService";
+import { policyService } from "../../services/policyService";
+import { Policy } from "../../types/policy";
 import { CustomerManagement } from "../ui/customer/CustomerManagement";
 import ConfirmationPopup from "../ui/popup/ConfirmationPopup";
 import { AddCustomer, CustomerFormData } from "../ui/customer/AddCustomer";
