+++ src/services/policyService.ts
@@ -1,27 +1,5 @@
 import { ApiService } from "./api";
-
-export interface Policy {
-  id: string;
-  policyNumber: string;
-  productName: string;
-  startDate: string;
-  endDate: string;
-  premium: number;
-  coverage: string;
-  renewalStatus: "ACTIVE" | "INACTIVE" | "PENDING";
-  application: {
-    id: string;
-    product: {
-      id: string;
-      name: string;
-      description: string;
-    };
-    status: string;
-    createdAt: string;
-    updatedAt: string;
-  };
-}
-
+import { Policy } from "../types/policy";
 class PolicyService {
   private apiService: ApiService;
